Implementará area de inserção de produtos
